Fix side effect of previous fix.
#26 moved the responsibility of executing scepter skills from possesion of the item to possession of the skilldef.
TreebotFireFruitSeed2 places a check in a global code path so the result from the query for the skilldef now needs to be null-checked since characters without a corresponding Skill Slot can run it on some circumstances. The null conditional operator is ok here because GetSkill returns an explicit csharp `null` when the GenericSkill is missing.