Extra checks for upgraded GenericSkills.
This fixes a case where states targeted by scepter can be re-used by the
game for different bodies,which can break assumptions that are normally
always true (such as a skill existing in the slot it's in).
The specific instance of this bug that was found was Incinerator Drones
using Artificer's flamethrower state for their attack.